chore: add files missed from PR 2083
